WIP-MEI: Robuster encoding detection in SubConverter.parseFile()
On POSIX systems (i.e., everything but Windows) the "file" command is used to
help determine the character encoding of a file about to be imported, ignoring
any declared encoding within the file. On non-POSIX systems, parseFile() tries
first with UTF-8 then UTF-16 before allowing a failure. UTF-16 is used by the
Sibelius-to-MEI exporter, so it will appear relatively more often on Windows(!)